+++ spec/ca_testing/drivers/v4/remote_spec.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
-RSpec.describe CaTesting::Drivers::V4::Local do
+RSpec.describe CaTesting::Drivers::V4::Remote do
   before do
     Capybara.default_driver = :selenium
     described_class.new(browser).register
@@ -14,11 +14,11 @@ RSpec.describe CaTesting::Drivers::V4::Local do
       let(:browser) { :chrome }
 
       it "has correct top level properties" do
-        expect(options.keys).to eq(%i[browser clear_local_storage clear_session_storage service capabilities])
+        expect(options.keys).to eq(%i[browser clear_local_storage clear_session_storage capabilities url])
       end
 
       it "has correct desired capabilities" do
-        expect(options[:capabilities].first.as_json).to eq({})
+        expect(options[:capabilities].first.as_json).to eq({ "browserName" => "chrome" })
       end
 
       it "has correct browser options" do
@@ -36,11 +36,11 @@ RSpec.describe CaTesting::Drivers::V4::Local do
       let(:browser) { :firefox }
 
       it "has correct top level properties" do
-        expect(options.keys).to eq(%i[browser clear_local_storage clear_session_storage service capabilities])
+        expect(options.keys).to eq(%i[browser clear_local_storage clear_session_storage capabilities url])
       end
 
       it "has correct desired capabilities" do
-        expect(options[:capabilities].first.as_json).to eq({})
+        expect(options[:capabilities].first.as_json).to eq({ "browserName" => "firefox" })
       end
 
       it "has correct browser options" do
@@ -54,32 +54,11 @@ RSpec.describe CaTesting::Drivers::V4::Local do
       end
     end
 
-    context "for internet explorer" do
-      let(:browser) { :internet_explorer }
-
-      it "has correct top level properties" do
-        expect(options.keys).to eq(%i[browser clear_local_storage clear_session_storage service capabilities])
-      end
-
-      it "has correct desired capabilities" do
-        expect(options[:capabilities].first.as_json).to eq({})
-      end
-
-      it "has correct (Modified), browser options" do
-        expect(options[:capabilities].last.as_json)
-          .to eq(
-            {
-              "se:ieOptions" => { "enablePersistentHover" => true, "nativeEvents" => true }
-            }
-          )
-      end
-    end
-
     context "for an unsupported browser" do
       let(:browser) { :foo }
 
       it "doesn't work if the browser is not one of the supported browsers" do
-        expect { options }.to raise_error(NoMethodError)
+        expect { options }.to raise_error(ArgumentError)
       end
     end
   end